spike(roads): Normalization Algorithm
* Rename the Distance normalizer to `RoadSegment` as that's
  more relevant to what we're trying to achieve.
* Make the RoadSegment normalizer work via an instance of
  the class - this allows us to memoize commonly calculated
  values.
